fix(summary): only request progress for authenticated users
